requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.28.72
botocore==1.31.72
CacheControl==0.13.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.6
django-3-jet==1.0.8
django-cors-headers==4.3.0
django-crontab==0.7.1
django-jet==1.0.8
django-redis==5.4.0
django-suit==0.2.25
djangorestframework==3.14.0
dnspython==2.6.1
eventlet==0.36.1
firebase-admin==6.3.0
gevent==24.2.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.6
msgpack==1.0.7
mysql-connector==2.2.9
mysql-connector-python==8.3.0
packaging==23.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.3
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
structlog==23.2.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.1
